cta {mcprofile}	R Documentation

Cell transformation assay dataset

Description

Balb//c 3T3 cells are treated with different concentrations of a carcinogen. Cells treated with a carcinogen do not stop proliferation. Number of foci (cell accumulations) are counted for 10 replicates per concentration level.

Usage

cta

Format

A data frame with 80 observations on the following 2 variables.

conc

a numeric vector denoting the concentration levels

foci

a numeric vector with the number of foci

References

Thomas C (2008): ECVAM data

[Package mcprofile version 1.0-0 Index]