summary.mcprofile {mcprofile}	R Documentation

Multiple Testing of General Hypotheses

Description

Multiple contrast testing based on signed root deviance profiles.

Usage

## S3 method for class 'mcprofile'
summary(object, margin = 0, adjust = "single-step",
  alternative = c("two.sided", "less", "greater"), ...)

Arguments

object
an object of class mcprofile

margin
test margin, specifying the right hand side of the hypotheses.

adjust
a character string specifying the adjustment for multiplicity. "single-step" controlling the FWER utilizing a multivariate normal- or t-distribution; "none" for comparison-wise error rate, or any other method provided by p.adjust.

alternative
a character string specifying the alternative hypothesis.

...
...

Value

An object of class mcpSummary

See Also

mcprofile, summary.glht

[Package mcprofile version 1.0-1 Index]